Orca: fix np 1.24 type deprecations (#7591)
* test(action.yml): rm numpy version require

* fix(pytorch-test): use np.float32

* fix(utils.py): dtype=object

* fix(run-pytests-ray.sh): add mxnet

* fix(utils.py): perf predictions

* perf(utils.py): change for loop

* fix(utils.py): typo

* fix(utils.py): revert

* style(utils.py): rename

* docs(utils.py): add comments

* fix(run-pytests-ray.sh): add TODO

* perf(utils.py): change comments

+++ python/orca/test/bigdl/orca/learn/ray/pytorch/test_estimator_pyspark_backend.py
@@ -236,7 +236,7 @@ class TestPyTorchEstimator(TestCase):
         sc = init_nncontext()
         spark = SparkSession.builder.getOrCreate()
         rdd = sc.range(0, 100)
-        data = rdd.map(lambda x: (np.random.randn(50).astype(np.float).tolist(),
+        data = rdd.map(lambda x: (np.random.randn(50).astype(np.float32).tolist(),
                                   [float(np.random.randint(0, 2, size=()))])
                        )
         schema = StructType([
@@ -283,7 +283,7 @@ class TestPyTorchEstimator(TestCase):
         spark = SparkSession.builder.getOrCreate()
         rdd = sc.range(0, 100)
         epochs = 2
-        data = rdd.map(lambda x: (np.random.randn(50).astype(np.float).tolist(),
+        data = rdd.map(lambda x: (np.random.randn(50).astype(np.float32).tolist(),
                                   [float(np.random.randint(0, 2, size=()))])
                        )
         schema = StructType([
